feat: refactor code #1
